% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohc.R
\docType{data}
\name{ohc}
\alias{ohc}
\title{One Horse Creek, New Zealand, Macroinvertebrate data}
\format{
## `ohc`
A vector of 1,809 body size observations:
}
\source{
<https://onlinelibrary.wiley.com/doi/abs/10.1111/fwb.13196>
}
\usage{
ohc
}
\description{
This is a vector of body mass (mg) observations from One Horse Creek in New Zealand. The original study surveyed streams across an acid mine drainage gradient. The body sizes in this data were aggregated from three surber samples. All individuals were measured and their body mass was estimated from published length-weight regressions.
}
\details{
This work was published in Freshwater Biology by Pomeranz et al. 2018 <https://onlinelibrary.wiley.com/doi/abs/10.1111/fwb.13196> and the full data set is available from the Dryad Digital Repository: <https://doi.org/10.5061/dryad.v6g985s>
}
\keyword{datasets}
