% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFSCarISD-package.R
\docType{package}
\name{UFSCarISD-package}
\alias{UFSCarISD}
\alias{UFSCarISD-package}
\title{UFSCarISD: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jpomz.github.io/UFSCarISD/}
}

}
\keyword{internal}
