% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLOU.R
\docType{data}
\name{WLOU}
\alias{WLOU}
\title{West St. Louis Creek}
\format{
## `WLOU`
A data frame with 4,558 rows and 4 columns:
\describe{
  \item{site_date}{This is a combination of the site name and the date it was collected}
  \item{organism_group}{Character value indicating either "Fish" or "Invertebrates"}
  \item{body_mass}{Weight, in mg drymass estimated from individual body lengths}
  \item{count_m2}{Estimated count of individuals on a per meter squared scale}
  ...
}
}
\source{
<https://www.biorxiv.org/content/10.1101/2024.01.09.574822v1>
}
\usage{
WLOU
}
\description{
Community data collected in September 2021 from West St. Louis Creek in Colorado, USA. This site is abbreviated as WLOU and is a part of the National Ecological Observatory Network (NEON). More information about NEON can be found at https://www.neonscience.org/
}
\details{
This data is an example of when to use the `negLL.PLB.counts` in the `sizeSpectra::calcLike()` function.

This data has been modified from NEON data products for macroinvertebrates (DP1.20120.001) and for fish (DP1.20107.001). The collections were not necessarily made on the exact same day.
}
\keyword{datasets}
