% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFSCarISD-package.R
\docType{package}
\name{UFSCarISD-package}
\alias{UFSCarISD}
\alias{UFSCarISD-package}
\title{UFSCarISD: Tutorial on using the "sizeSpectra" Package.}
\description{
This package was built for a short course on using the "sizeSpectra" package. The short course will be delivered to UFSCar in Sao Carlos Brazil on July 6th 2024. This package has an associated web site for participants to use during and after the short course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jpomz.github.io/UFSCarISD/}
}

}
\keyword{internal}
